move constants file into a dedicated folder, so we can scope variable injection to this file
